project_settings/requirements.txt
amqp==1.4.9
anyjson==0.3.3
apns2==0.2.1
appdirs==1.4.3
arrow==0.10.0
attrs==16.3.0
Babel==2.2.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
blessed==1.14.2
celery==3.1.24
cffi==1.10.0
coreapi==2.0.0
cryptography==1.7.1
decorator==4.0.10
diff-match-patch==20121119
Django==1.8.5
django-admin-rangefilter==0.2.1
django-ajax-selects==1.6.0
django-allauth==0.26.1
django-annoying==0.10.3
django-appconf==1.0.1
django-autocomplete-light==3.2.1
django-autoslug==1.9.3
django-braces==1.9.0
django-cachalot==1.2.1
django-celery==3.1.17
django-common-helpers==0.8.0
django-compressor==2.0
django-crispy-forms==1.6.0
django-cron==0.4.6
django-datatables-view==1.13.0
django-daterange-filter==1.3.0
django-debug-toolbar==1.8
django-easy-select2==1.3.3
-e git://github.com/mvpdev/django-eav.git@86098dc1cd602be8cf2e7fd8fbbc34a6e6b2dd53#egg=django_eav
django-extra-views==0.6.4
django-fcm==0.0.8
django-filter==0.15.3
django-fsm==2.6.0
django-grappelli==2.7.3
django-gravatar2==1.4.0
django-haystack==2.4.1
django-imagefit==0.4
django-imagekit==3.3
django-impersonate==1.0.1
django-import-export==0.4.5
django-jsonfield==0.9.15
django-last-modified==0.1
django-livefield==2.8.0
django-loginas==0.2.0
django-model-utils==3.0.0
django-multiselectfield==0.1.3
django-notifier==0.7
django-oscar==1.2.1
django-oscar-api==1.0.3
django-picklefield==0.3.2
django-push-notifications==1.4.1
django-pyodbc==1.1.0
django-pyodbc-azure==1.8.1.0
django-q==0.8.0
django-redis==4.6.0
django-redis-cache==1.7.1
django-rest-auth==0.8.1
django-rest-multitoken==0.0.0
django-rest-swagger==2.0.5
django-reversion==1.10.0
django-seed==0.1.3
django-select-multiple-field==0.4.2
django-silver==0+unknown
django-simple-history==1.7.0
django-su==0.5.2
django-tables2==1.0.7
django-treebeard==4.0
django-userlog==0.2
django-versatileimagefield==1.7.0
django-widget-tweaks==1.4.1
django-xhtml2pdf==0.0.3
django-xmpp==0.3.2
djangorestframework==3.3.2
djangorestframework-bulk==0.2.1
djangorestframework-jsonp==1.0.0
dnspython==1.12.0
docopt==0.6.2
drf-extensions==0.2.8
drf-nested-routers==0.11.1
drf-tracking==0.2.5
drfdocs==0.0.11
enum34==1.0.4
factory-boy==2.6.1
fake-factory==0.5.2
fakeredis==0.7.0
fcm-django==0.2.2
funcsigs==0.4
furl==1.0.0
future==0.16.0
gdata==2.0.18
gevent==1.1.2
google-api-python-client==1.4.2
greenlet==0.4.10
grequests==0.3.0
h2==2.6.2
hiredis==0.2.0
hpack==3.0.0
html5lib==0.999
httplib2==0.9.2
hyper==0.7.0
hyperframe==3.2.0
idna==2.0
ipaddress==1.0.14
ipython-genutils==0.1.0
itypes==1.1.0
kombu==3.0.37
Markdown==2.6.2
mock==1.3.0
MySQL-python==1.2.5
ndg-httpsclient==0.4.2
ngxtop==0.0.2
oauth2client==1.5.1
oauthlib==1.0.1
olefile==0.44
openapi-codec==1.0.0
orderedmultidict==0.7.11
packaging==16.8
pathlib2==2.1.0
pbr==1.8.1
pexpect==4.2.0
phonenumbers==7.2.8
pickleshare==0.7.3
pilkit==1.1.13
Pillow==4.0.0
prompt-toolkit==1.0.5
protobuf==3.0.0b2.post2
ptyprocess==0.5.1
purl==1.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycountry==17.5.14
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pyfcm==1.0.1
Pygments==2.1.3
PyJWT==1.5.0
pyodbc==4.0.11
pyOpenSSL==0.15.1
pyparsing==2.2.0
PyPDF2==1.26.0
python-axolotl==0.1.34
python-axolotl-curve25519==0.1
python-dateutil==2.5.2
python-memcached==1.58
python-openid==2.2.5
pytz==2016.7
pyvat==1.3.1
pywhatsapp==0.0.5
rcssmin==1.0.6
redis==2.10.5
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.5.0
rjsmin==1.0.12
rsa==3.2.3
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.10.0
sleekxmpp==1.3.1
sorl-thumbnail==12.3
SQLAlchemy==1.1.0
sqlparse==0.2.3
tablib==0.11.2
tabulate==0.7.5
traitlets==4.2.2
Unidecode==0.4.19
UNKNOWN==0.0.0
uritemplate==0.6
uWSGI==2.0.13.1
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.10.4
xhtml2pdf==0.0.6
yowsup2==2.4.102
